// src/api/models/error-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { DetailedErrorModel } from './detailed-error-model';
// May contain unused imports in some cases
// @ts-ignore
import { ErrorModel } from './error-model';
// May contain unused imports in some cases
// @ts-ignore
import { WarningModel } from './warning-model';

/**
 * 
 * @export
 * @interface ErrorResponse
 */
export interface ErrorResponse {
    /**
     * 
     * @type {string}
     * @memberof ErrorResponse
     */
    'code'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ErrorResponse
     */
    'message'?: string | null;
    /**
     * 
     * @type {Array<ErrorModel>}
     * @memberof ErrorResponse
     */
    'messages'?: Array<ErrorModel> | null;
    /**
     * 
     * @type {Array<DetailedErrorModel>}
     * @memberof ErrorResponse
     */
    'detailedMessages'?: Array<DetailedErrorModel> | null;
    /**
     * 
     * @type {Array<WarningModel>}
     * @memberof ErrorResponse
     */
    'warningMessages'?: Array<WarningModel> | null;
}