// src/api/models/role-profile-model-paged.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { RoleProfileModel } from './role-profile-model';

/**
 * 
 * @export
 * @interface RoleProfileModelPaged
 */
export interface RoleProfileModelPaged {
    /**
     * 
     * @type {number}
     * @memberof RoleProfileModelPaged
     */
    'totalItems'?: number;
    /**
     * 
     * @type {number}
     * @memberof RoleProfileModelPaged
     */
    'page'?: number;
    /**
     * 
     * @type {number}
     * @memberof RoleProfileModelPaged
     */
    'totalPages'?: number;
    /**
     * 
     * @type {Array<RoleProfileModel>}
     * @memberof RoleProfileModelPaged
     */
    'items'?: Array<RoleProfileModel> | null;
    /**
     * 
     * @type {number}
     * @memberof RoleProfileModelPaged
     */
    'itemsCount'?: number | null;
}